refactoring quick deploy code to make naming more consinsent
