+++ GooglePlayInstant/Editor/PlayInstantEditorMenu.cs
@@ -14,6 +14,7 @@
 
 using UnityEditor;
 using UnityEngine;
+using GooglePlayInstant.Editor.QuickDeploy;
 
 namespace GooglePlayInstant.Editor
 {
@@ -71,31 +72,31 @@ namespace GooglePlayInstant.Editor
         [MenuItem("PlayInstant/Quick Deploy/AssetBundle Creation...", false, 300)]
         private static void AssetBundleCreationSettings()
         {
-            PlayInstantQuickDeployWindow.ShowWindow(PlayInstantQuickDeployWindow.ToolBarSelectedButton.CreateBundle);
+            QuickDeployWindow.ShowWindow(QuickDeployWindow.ToolBarSelectedButton.CreateBundle);
         }
 
         [MenuItem("PlayInstant/Quick Deploy/AssetBundle Deployment...", false, 301)]
         private static void AssetBundleDeploymentSettings()
         {
-            PlayInstantQuickDeployWindow.ShowWindow(PlayInstantQuickDeployWindow.ToolBarSelectedButton.DeployBundle);
+            QuickDeployWindow.ShowWindow(QuickDeployWindow.ToolBarSelectedButton.DeployBundle);
         }
 
         [MenuItem("PlayInstant/Quick Deploy/AssetBundle Verification...", false, 302)]
         private static void AssetBundleVerificationSettings()
         {
-            PlayInstantQuickDeployWindow.ShowWindow(PlayInstantQuickDeployWindow.ToolBarSelectedButton.VerifyBundle);
+            QuickDeployWindow.ShowWindow(QuickDeployWindow.ToolBarSelectedButton.VerifyBundle);
         }
 
         [MenuItem("PlayInstant/Quick Deploy/Loading Screen...", false, 303)]
         private static void LoadingScreenSettings()
         {
-            PlayInstantQuickDeployWindow.ShowWindow(PlayInstantQuickDeployWindow.ToolBarSelectedButton.LoadingScreen);
+            QuickDeployWindow.ShowWindow(QuickDeployWindow.ToolBarSelectedButton.LoadingScreen);
         }
 
         [MenuItem("PlayInstant/Quick Deploy/Build APK...", false, 304)]
         private static void BuildApkSettings()
         {
-            PlayInstantQuickDeployWindow.ShowWindow(PlayInstantQuickDeployWindow.ToolBarSelectedButton.Build);
+            QuickDeployWindow.ShowWindow(QuickDeployWindow.ToolBarSelectedButton.Build);
         }
 
         [MenuItem("PlayInstant/Build and Run #%r", false, 400)]
