+++ GooglePlayInstant/Editor/QuickDeploy/AssetBundleBrowserClient.cs
@@ -19,7 +19,7 @@ using System.Text.RegularExpressions;
 using UnityEditor;
 using UnityEngine;
 
-namespace GooglePlayInstant.Editor
+namespace GooglePlayInstant.Editor.QuickDeploy
 {
     /// <summary>
     /// A static class responsible for interactions with the Asset Bundle Browser
