+++ GooglePlayInstant/Editor/QuickDeploy/AssetBundleVerifierWindow.cs
@@ -16,7 +16,7 @@ using UnityEditor;
 using UnityEngine;
 using UnityEngine.Networking;
 
-namespace GooglePlayInstant.Editor
+namespace GooglePlayInstant.Editor.QuickDeploy
 {
     /// <summary>
     /// Window that verifies AssetBundles from given URLs.
